+++ test/components/component_test.rb
@@ -39,6 +39,7 @@ class PrimerComponentTest < Minitest::Test
     [Primer::FlashComponent, {}],
     [Primer::FlexItemComponent, { flex_auto: true }],
     [Primer::HeadingComponent, {}],
+    [Primer::HiddenTextExpander, {}],
     [Primer::LabelComponent, { title: "Hello!" }],
     [Primer::LayoutComponent, {}],
     [Primer::LinkComponent, { href: "https://www.google.com" }],